refactor: BaseMapViewModel로 지도 공통 로직 분리 후 필요에 따라 상속 받도록 리팩토링
