+++ river/Starbucks_Clone/Starbucks_Clone/Sources/Screen/Map/Protocol/MapControllable.swift
@@ -14,4 +14,7 @@ protocol MapControllable: Observable, AnyObject {
     var storeList: [StoreList] { get }
     
     var nearbyStores: [StoreList] { get }
+    
+    /// 스타벅스 위치 정보 로드하는 함수(GeoJson 파싱)
+    func loadStarbucksStores()
 }
